% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA24204139.R
\name{complete.design}
\alias{complete.design}
\title{Complete Experimental Design}
\usage{
complete.design(varlist, numbering = TRUE, size = 1)
}
\arguments{
\item{varlist}{A list where each element represents a factor.
Each element can be:
\itemize{
\item A positive integer (representing the number of levels for a factor).
\item A vector (representing the specific levels of a factor).
\item A data frame (representing a complex factor with multiple columns).
}}

\item{numbering}{Logical, whether to include experiment numbers (default = TRUE).}

\item{size}{Integer, representing the number of repetitions for each combination (default = 1).}
}
\value{
A data frame containing the full factorial design.
}
\description{
This function generates a complete factorial experimental design.
}
\examples{
complete.design(list(f1 = 3, f2 = c("a", "b"), f3 = 4))
complete.design(list(f1 = 3, f2 = c("a", "b"),
  f3 = data.frame(fa = c(0.01, 0.05, 0.1), fb = c("x", "x", "y"))), size = 2)
}
